Refactored PMXFile PMXParser

+++ src/common/Locale.cpp
@@ -10,7 +10,7 @@
 namespace mm::Locale
 {
 
-String UTF16LEToUTF8(const char* s, size_t len)
+String UTF16LEToUTF8(const char* s, i32 len)
 {
 	static std::wstring_convert<std::codecvt_utf8_utf16<char16_t, 0x10ffffu, std::codecvt_mode::little_endian>, char16_t> conv;
 	std::string str(conv.to_bytes(std::u16string((char16_t*)s, (char16_t*)&s[len])));
@@ -19,7 +19,7 @@ String UTF16LEToUTF8(const char* s, size_t len)
 
 /* Include null terminator */
 /* https://gist.github.com/kilfu0701/e279e35372066ae1832850c438d5611e */
-String ShiftJISToUTF8(const char* s, size_t len)
+String ShiftJISToUTF8(const char* s, i32 len)
 {
 	std::string sjfstr(s, s + len);
 	icu::UnicodeString src(sjfstr.c_str(), "shift_jis");
@@ -31,7 +31,7 @@ String ShiftJISToUTF8(const char* s, size_t len)
 	return std::string(result.begin(), result.end() - 1).append("");
 }
 
-String UTF8ToShiftJIS(const char* s, size_t len)
+String UTF8ToShiftJIS(const char* s, i32 len)
 {
 	icu::UnicodeString src(s, "utf8");
 	int length = src.extract(0, src.length(), NULL, "shift_jis");
