+++ src/common/Locale.hpp
@@ -1,13 +1,14 @@
 #pragma once
 
 #include "Container/String.hpp"
+#include "Base.hpp"
 
 namespace mm::Locale
 {
 
-String UTF16LEToUTF8(const char* s, size_t len);
-String ShiftJISToUTF8(const char* s, size_t len);
-String UTF8ToShiftJIS(const char* s, size_t len);
+String UTF16LEToUTF8(const char* s, i32 len);
+String ShiftJISToUTF8(const char* s, i32 len);
+String UTF8ToShiftJIS(const char* s, i32 len);
 
 }
